Project Backend/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
email-validator==2.1.1
sqlalchemy==2.0.38
passlib[bcrypt]==1.7.4
bcrypt==4.1.2